+++ veramo-scenario/src/ethr-registry-lib/test-samples/add-service.ts
@@ -1,0 +1,25 @@
+import { agent } from './veramo/setup'
+
+//***future work : need monitoring documentation***// 
+async function main() {
+  const option = {
+    did: "did:ethr:rinkeby:0x03ba1cf1fa7ebb6ea862ee2104423d3aa790e59e4d61dd487ea088108fca1a5dfa",
+    service:{
+        description: "alummi",
+        id: "#prove",
+        type: "prove",
+        serviceEndpoint:"www.example.com"
+    },
+    options:{
+      provider: 'did:ethr:rinkeby'
+    }
+  }
+  const identity = await agent.didManagerAddService(option);
+  console.log(`New identity created`)
+  console.log(identity)
+}
+
+main().catch(err=>{
+    console.log("---fail---")
+    console.log(err);
+});